Add favicon support to HTML files; update ResourceRepository to use Integer for ID; refactor ResourceItem constructor and getter

+++ resourcify-backend/resourcify-backend/src/main/java/com/resourcify/resourcify_backend/repository/ResourceRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ResourceRepository extends JpaRepository<ResourceItem, Long> {
+public interface ResourceRepository extends JpaRepository<ResourceItem, Integer> {
 
     @Query("SELECT DISTINCT r.location FROM ResourceItem r")
     List<String> findAllDistinctLocations();
